Change default timestamp to 64 bits in all webrtc directories.
BUG=
R=pbos@webrtc.org, pthatcher@webrtc.org, solenberg@webrtc.org

Review URL: https://codereview.webrtc.org/1835053002 .

Cr-Commit-Position: refs/heads/master@{#12646}

+++ webrtc/api/peerconnectionendtoend_unittest.cc
@@ -38,7 +38,7 @@ using webrtc::PeerConnectionInterface;
 
 namespace {
 
-const size_t kMaxWait = 10000;
+const int kMaxWait = 10000;
 
 }  // namespace
