+++ webrtc/pc/currentspeakermonitor.h
@@ -54,7 +54,7 @@ class CurrentSpeakerMonitor : public sigslot::has_slots<> {
   // Used by tests.  Note that the actual minimum time between switches
   // enforced by the monitor will be the given value plus or minus the
   // resolution of the system clock.
-  void set_min_time_between_switches(uint32_t min_time_between_switches);
+  void set_min_time_between_switches(int min_time_between_switches);
 
   // This is fired when the current speaker changes, and provides his audio
   // SSRC.  This only fires after the audio monitor on the underlying
@@ -86,8 +86,8 @@ class CurrentSpeakerMonitor : public sigslot::has_slots<> {
   uint32_t current_speaker_ssrc_;
   // To prevent overswitching, switching is disabled for some time after a
   // switch is made.  This gives us the earliest time a switch is permitted.
-  uint32_t earliest_permitted_switch_time_;
-  uint32_t min_time_between_switches_;
+  int64_t earliest_permitted_switch_time_;
+  int min_time_between_switches_;
 };
 
 }  // namespace cricket
